requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers